move: 팔로우 버튼 수정  (#26)
* feat:기초 디자인설정

* style:간단한 style 수정

* style: 화면 상단으로 가는 버튼 생성. 무한스크롤은 나중에 할듯

* feature/home/style 팔레트활용 컬러 넣음

* feat/home/style/add 버튼 위치 옮김-수파베이스활용 더미데이터 써봄

* feat/Home/add: FetchData.jsx에서 GameRankFetchData.jsx로 파일명 변경 및 페이지네이션 기능 구현추가

* feat/home/style:펫치 데이타 카드 색 수정 및 랭크번호 삽입

* feat:chore GameRankFetchData.jsx파일 위치이동

* feat:검색기능과 장르별 토글 드롭다운 버튼 기능생성 및 구현

* style: homepage스타일 컴포넌트 분리 및 기본 디자인 수정

* style:상단 카드크기 수정

* style:상단 카드 컨테이너 크기 대폭 축소

* style: 커뮤니티 박스 스타일조정

* style:css대폭수정

* style:css변경

* add:페이지네이션,하단 글 검색기능

* style:font size rem으로 조정

* style rem에서 px로 다시 바꿨습니다

* chore쓸데없는 라이브러리 지움

* add:env.local추가 수파베이스 연동

* fix:dev호환성

* update:follow 기능 구현

---------
